% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jump_matrix.R
\name{create_jump_matrix}
\alias{create_jump_matrix}
\title{Create jump matrix}
\usage{
create_jump_matrix(multiplex)
}
\arguments{
\item{multiplex}{a multiplex}
}
\value{
jump_mat a jump matrix
}
\description{
Create jump matrix
}
