% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layer_transition_matrix.R
\name{create_layer_transition_matrix}
\alias{create_layer_transition_matrix}
\title{Create jump matrix containing the probabilities to switch from one omics to another}
\usage{
create_layer_transition_matrix(omics_list)
}
\arguments{
\item{omics_list}{A list of DataFrames containings omics data, each is a DataFrame with three columns: source, destination, weight}
}
\value{
A Matrix containing the probabilities to switch from one omics to another
}
\description{
Create jump matrix containing the probabilities to switch from one omics to another
}
