% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_mtx-data.R
\docType{data}
\name{GBM_mtx}
\alias{GBM_mtx}
\title{Phenotype data of GBM use case}
\format{
The matrices included in GBM_mtx are:
  \describe{
  \item{GliomaCNV_norm}{A matrix representing copy number variations (CNV) data for GBM samples, with genes as features on rows}
  \item{GliomaMethylation_norm}{A matrix representing DNA methylation data for GBM samples, with individual CpG sites as features on rows}
  \item{GliomaExpression_norm}{A matrix representing gene expression data for GBM samples, with genes as features on rows}
}
}
\usage{
data(GBM_mtx)
}
\description{
GBM_mtx is a list of matrices containing molecular data on glioblastoma (GBM) samples.
Specifically, each matrix in the list contains features on rows and samples on columns,
and represents different molecular data types.
}
\examples{
data(GBM_mtx)
}
\keyword{datasets}
