% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_mat_M.R
\name{gen_sim_mat_M}
\alias{gen_sim_mat_M}
\title{Create jump matrix}
\usage{
gen_sim_mat_M(
  network,
  tau = NA,
  restart = 0.7,
  delta = 0.5,
  cond_jump = NULL,
  cores = 1
)
}
\arguments{
\item{network}{a network}

\item{tau}{tau}

\item{restart}{restart}

\item{delta}{delta}

\item{cond_jump}{cond_jump}

\item{cores}{number of threads launched}
}
\value{
RWRM_similarity
}
\description{
Create jump matrix
}
