% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection_matrices.R
\name{get_intersection_matrices}
\alias{get_intersection_matrices}
\title{Get intersection between cancer matrices}
\usage{
get_intersection_matrices(omics_list)
}
\arguments{
\item{omics_list}{a list of omics}
}
\description{
Get intersection between cancer matrices
}
