% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsne_embedding.R
\name{get_tsne_embedding}
\alias{get_tsne_embedding}
\title{TSNE embedding of matrix}
\usage{
get_tsne_embedding(matrix, embedding_size, perplexity, max_iter, num_threads)
}
\arguments{
\item{matrix}{a squared n n x n numeric matrix with values in the range 0-1, with observations on columns}

\item{embedding_size}{size of the output embedding}

\item{perplexity}{perplexity parameter}

\item{max_iter}{number of iterations}

\item{num_threads}{number of threads for parallelization. It has to be positive integer or 0, if all available cores must be used. If it is equal to 1, no parallelization is not performed}
}
\value{
a \emph{embedding_size} x n matrix having observations on columns
}
\description{
TSNE embedding of matrix
}
