% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_star_net.R
\name{k_star_net}
\alias{k_star_net}
\title{Do k_star}
\usage{
k_star_net(mat, distFun, sparsity = 1, knn = 25, cores = 1)
}
\arguments{
\item{mat}{an omics matrix}

\item{distFun}{a distance function}

\item{sparsity}{a positive real}

\item{knn}{an integer}

\item{cores}{number of threads}
}
\value{
tibble colomuns : source, destination, weight
}
\description{
Do k_star
}
