% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_omics.R
\name{normalize_omics}
\alias{normalize_omics}
\title{normalize_omics takes in input a Matrix that has to be centered and scaled along the columns}
\usage{
normalize_omics(mtx)
}
\arguments{
\item{mtx}{A numeric Matrix representing a particular omics, rows are genes and columns are samples}
}
\value{
An object of the same type as mtx. The matrix mtx is normalize along the columns
}
\description{
normalize_omics takes in input a Matrix that has to be centered and scaled along the columns
}
