% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_matrix.R
\name{plot_2D_matrix}
\alias{plot_2D_matrix}
\title{Plot bidimensional matrix}
\usage{
plot_2D_matrix(
  coord,
  nodes_anno,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  interactive = TRUE,
  wo_legend = FALSE,
  title = "",
  shapes = NULL
)
}
\arguments{
\item{coord}{A DataFrame with samples on columns that has to be plotted}

\item{nodes_anno}{Annotation DataFrame with all the information for each sample}

\item{id_name}{String for identification of sample in nodes_anno}

\item{id_anno_color}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific color for each case}

\item{id_anno_shape}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific shape for each case, it must be associated to a discrete variable}

\item{interactive}{A boolean flag, if TRUE returns an interactive plot}

\item{wo_legend}{A boolean flag, if TRUE returns the plot without legend}

\item{title}{Title of the plot}

\item{shapes}{It is possible to give the shapes list for plotting, id_anno_shape is not NA}
}
\value{
A plot
}
\description{
Plot bidimensional matrix
}
