% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Plot omics network with annotations}
\usage{
plot_net(
  edgeList,
  nodes_anno,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  title = "",
  html = FALSE,
  wo_legend = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{edgeList}{a data frame containing the edge list of a network in the first two columns containing node ids.}

\item{nodes_anno}{annotation data frame with nodes on rows and annotations on columns}

\item{id_name}{name of the column in \emph{nodes_anno} containing unique sample ids}

\item{id_anno_color}{name of the column in \emph{nodes_anno} whose values are mapped to colors to observations. Notice that colors can be recycled it eh number of distinct values exceed available shapes}

\item{id_anno_shape}{name of the column in \emph{nodes_anno} whose values are mapped to shapes to observations. Notice that shapes can be recycled it eh number of distinct values exceed available shapes}

\item{title}{title of plot}

\item{html}{logical, if TRUE the plot is produced in html format}

\item{wo_legend}{logical, if TRUE returns the plot without legend}

\item{interactive}{logical, if TRUE returns an interactive plot}
}
\description{
Plot omics network with annotations
}
