% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parallel_umap.R
\name{plot_parallel_umap}
\alias{plot_parallel_umap}
\title{Plot parallel umap}
\usage{
plot_parallel_umap(
  matrix,
  nodes_anno,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  interactive = TRUE,
  wo_legend = FALSE,
  title = "",
  n_neighbors = 15,
  n_threads = NULL,
  n_sgd_threads = 0,
  grain_size = 1
)
}
\arguments{
\item{matrix}{A Matrix with samples on columns that has to be plotted}

\item{nodes_anno}{Annotation DataFrame with all the information for each sample}

\item{id_name}{String for identification of sample in nodes_anno}

\item{id_anno_color}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific color for each case}

\item{id_anno_shape}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific shape for each case, it must be associated to a discrete variable}

\item{interactive}{A boolean flag, if TRUE returns an interactive plot}

\item{wo_legend}{A boolean flag, if TRUE returns the plot without legend}

\item{title}{Title of the plot}

\item{n_neighbors}{The size of local neighborhood. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100}

\item{n_threads}{Number of thread, NULL if single core}

\item{n_sgd_threads}{Number of thread for stochastic gradient descent, if > 1 it will not be reproducible anyway}

\item{grain_size}{Minimum amount of work to do on each thread}
}
\value{
A plot
}
\description{
Plot parallel umap
}
