% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parallel_umap_db.R
\name{plot_parallel_umap_db}
\alias{plot_parallel_umap_db}
\title{Plot parallel umap db}
\usage{
plot_parallel_umap_db(
  matrix,
  k = 1,
  interactive = TRUE,
  n_neighbors = 15,
  n_threads = NULL,
  n_sgd_threads = 0,
  grain_size = 1
)
}
\arguments{
\item{matrix}{A DataFrame with samples on columns that has to be plotted}

\item{k}{Reachability distance, dbscan}

\item{interactive}{A boolean flag, it TRUE returns an interactive plot}

\item{n_neighbors}{The size of local neighborhood. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100}

\item{n_threads}{Number of thread, NULL if single core}

\item{n_sgd_threads}{Number of thread for stochastic gradient descent, if > 1 it will not be reproducible anyway}

\item{grain_size}{Minimum amount of work to do on each thread}
}
\description{
Plot parallel umap db
}
