% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_multiplex_network.R
\name{prune_multiplex_network}
\alias{prune_multiplex_network}
\title{Prune low weight edges from network}
\usage{
prune_multiplex_network(network, weight_remv = NULL)
}
\arguments{
\item{weight_remv}{a threshold value on the edge weight. Edges with weight less then \emph{weight_remv} will be removed from the network.}

\item{networka}{data frame containing the edge list of a network with two columns containing nodes ids and one column named weight containing edge weights}
}
\value{
data frame contianing pruned network as an edge list
}
\description{
Prune low weight edges from network
}
