% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_extra_weigth.R
\name{remove_extra_weigth}
\alias{remove_extra_weigth}
\title{Remove extra weigth}
\usage{
remove_extra_weigth(network, weight_remv = NULL)
}
\arguments{
\item{network}{a network with three columns}

\item{weight_remv}{the weight to be removed in the network}
}
\value{
network with weigths strictly less than weight_remv
}
\description{
Remove extra weigth
}
