% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zeros_cols.R
\name{remove_zeros_cols}
\alias{remove_zeros_cols}
\title{This function clean a Matrix removing the columns with all zeros}
\usage{
remove_zeros_cols(matrix)
}
\arguments{
\item{matrix}{A Matrix}
}
\value{
A Matrix of the same type of mtx without columns with all zeros
}
\description{
This function clean a Matrix removing the columns with all zeros
}
