% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jump_matrix.R
\name{create_jump_matrix}
\alias{create_jump_matrix}
\title{Create jump matrix containing the probabilities to switch from one omics to another}
\usage{
create_jump_matrix(multiplex)
}
\arguments{
\item{multiplex}{A multiplex, an object that comprises different omics. It is a DataFrame containing these information: EdgeType, source, target, weight}
}
\value{
A Matrix containing the probabilities to switch from one omics to another
}
\description{
Create jump matrix containing the probabilities to switch from one omics to another
}
