% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layer_transition_matrix.R
\name{create_layer_transition_matrix}
\alias{create_layer_transition_matrix}
\title{Create transition probabilities matrix}
\usage{
create_layer_transition_matrix(omics_list)
}
\arguments{
\item{omics_list}{a named list of data frames containing an omics multiplex network.
Each data frame in the list is required to have at least two columns containing nodes ids respectively named source and destination
and one column named weight containing edge weights.}
}
\value{
A squared n x n matrix (where n is the number of omics layers) containing  transition probabilities between omics layers.
}
\description{
Create transition probabilities matrix
}
