% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiplex.R
\name{create_multiplex}
\alias{create_multiplex}
\title{Create multiplex}
\usage{
create_multiplex(omics_list)
}
\arguments{
\item{omics_list}{a list of omics}
}
\value{
multiplex
}
\description{
Create multiplex
}
