% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiplex.R
\name{create_multiplex}
\alias{create_multiplex}
\title{Create a multiplex omic network}
\usage{
create_multiplex(omics_list, weighted = FALSE)
}
\arguments{
\item{omics_list}{a named list of data frames containing adjacency lists for different omics networks sharing node ids.
Each data frame in the list is required to have at least two columns containing nodes ids respectively named source and destination and one column named weight containing edge weights.}

\item{weighted}{logical. If TRUE then a weighted multiplex is returned, otherwise weights are all set to 1. FALSE by default.}
}
\value{
A data frame representing the edge list of a multiplex network with the following four columns: EdgeType , source, target, weight.
EdgeType is a character representing the layers in the multiplex, source and target represent the edges and contain node ids, weight contains edge weights in the multiplex network.
}
\description{
Create a multiplex omic network
}
