% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_mat_M.R
\name{gen_sim_mat_M}
\alias{gen_sim_mat_M}
\title{Generate similarity matrix}
\usage{
gen_sim_mat_M(
  network,
  tau = NA,
  restart = 0.7,
  delta = 0.5,
  cond_jump = NULL,
  cores = 1
)
}
\arguments{
\item{network}{a network}

\item{tau}{tau}

\item{restart}{A real in the range of 0-1, it is the probability to restart the algorithm in the starting point}

\item{delta}{delta}

\item{cond_jump}{cond_jump}

\item{cores}{Number of threads for Parallelization. It has to be positive integer. If it is equal to 1, no parallelization is not performed}
}
\value{
RWRM_similarity
}
\description{
Generate similarity matrix
}
