% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_mat_M.R
\name{gen_sim_mat_M}
\alias{gen_sim_mat_M}
\title{Generate similarity matrix
Adapted from:
https://github.com/LPioL/MultiVERSE/blob/master/RWR/Functions_RWRM.R
https://github.com/LPioL/MultiVERSE/blob/master/RWR/GenerateSimMatrix.R}
\usage{
gen_sim_mat_M(
  network,
  tau = NA,
  restart = 0.7,
  delta = 0.5,
  layer_transition = NULL,
  jump_neighborhood = FALSE,
  weighted_multiplex = FALSE,
  cores = 1
)
}
\arguments{
\item{network}{a network}

\item{tau}{tau}

\item{restart}{A real in the range of 0-1, it is the probability to restart the algorithm in the starting point}

\item{delta}{delta}

\item{layer_transition}{cond_jump}

\item{jump_neighborhood}{A boolean, if false, the nodes between omics will be connected if they are the same, if true, the nodes will be connected with itself and its neighborhood but in the other omics}

\item{weighted_multiplex}{A boolean, if true, the edge between omics will be weighted. It is considered only if jump_neighborhood is true}

\item{cores}{Number of threads for Parallelization. It has to be positive integer. If it is equal to 1, no parallelization is not performed}
}
\value{
RWRM_similarity
}
\description{
Research group:  Léo Pio-Lopez, Alberto Valdeolivas, Laurent Tichit, Élisabeth Remy, Anaïs Baudot
Publication:     https://arxiv.org/abs/2008.10085
}
