% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embedding.R
\name{get_embedding}
\alias{get_embedding}
\title{Multi-omics matrix embedding}
\usage{
get_embedding(matrix, embedding_size, num_steps = NA, cores = 20)
}
\arguments{
\item{matrix}{a n x n squared numeric similarity matrix with values in the range 0-1}

\item{embedding_size}{size of the output embedding}

\item{num_steps}{number of total epoches}

\item{cores}{number of threads for parallelization. It has to be positive integer. If it is equal to 1, no parallelization is performed}
}
\value{
a \emph{embedding_size} x n numerical matrix
}
\description{
Adapted from:
Research group:  Léo Pio-Lopez, Alberto Valdeolivas, Laurent Tichit, Élisabeth Remy, Anaïs Baudot
Publication:     https://arxiv.org/abs/2008.10085
}
