% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embedding.R
\name{get_embedding}
\alias{get_embedding}
\title{Create jump matrix}
\usage{
get_embedding(mtx, EMBED_DIMENSION, NUM_STEPS = 10^7, cores = 20)
}
\arguments{
\item{mtx}{a RWR matrix}

\item{EMBED_DIMENSION}{dimension of embedding}

\item{NUM_STEPS}{number of epoches}

\item{cores}{number of cores for parallelism}
}
\value{
embedding
}
\description{
Create jump matrix
}
