% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embedding.R
\name{get_embedding}
\alias{get_embedding}
\title{Compute an embedding of the multi-omics matrix}
\usage{
get_embedding(matrix, embedding_size, num_steps = 10^7, cores = 20)
}
\arguments{
\item{matrix}{A squared numeric matrix with values in the range 0-1, with samples on columns}

\item{embedding_size}{Size of the output embedding}

\item{num_steps}{Number of total epoches}

\item{cores}{Number of threads for Parallelization. It has to be positive integer. If it is equal to 1, no parallelization is not performed}
}
\value{
An embedding of the input Matrix: on the columns there are the samples, the number of rows are specified by embedding_size
}
\description{
Adapted from:
Research group:  Léo Pio-Lopez, Alberto Valdeolivas, Laurent Tichit, Élisabeth Remy, Anaïs Baudot
Publication:     https://arxiv.org/abs/2008.10085
}
