% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_umap_embedding.R
\name{get_parallel_umap_embedding}
\alias{get_parallel_umap_embedding}
\title{UMAP embedding of matrix using uwot umap}
\usage{
get_parallel_umap_embedding(
  matrix,
  embedding_size,
  n_neighbors = 15,
  n_threads = NULL,
  n_sgd_threads = 0,
  grain_size = 1,
  metric = "cosine"
)
}
\arguments{
\item{matrix}{a squared n n x n numeric matrix with values in the range 0-1, with observations on columns}

\item{embedding_size}{size of the output embedding}

\item{n_neighbors}{size of local neighborhood. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100}

\item{n_threads}{number of threads, NULL if single core}

\item{n_sgd_threads}{number of thread for stochastic gradient descent, if > 1 it will not be reproducible}

\item{grain_size}{minimum amount of work to do on each thread}

\item{metric}{distance metric to use with umap}
}
\value{
a \emph{embedding_size} x n matrix having observations on columns
}
\description{
UMAP embedding of matrix using uwot umap
}
