% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca_embedding.R
\name{get_pca_embedding}
\alias{get_pca_embedding}
\title{PCA embedding of matrix}
\usage{
get_pca_embedding(matrix, embedding_size)
}
\arguments{
\item{matrix}{a squared n n x n numeric matrix with values in the range 0-1, with observations on columns}

\item{embedding_size}{size of the output embedding}
}
\value{
a \emph{embedding_size} x n matrix having observations on columns
}
\description{
PCA embedding of matrix
}
