% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsne_embedding.R
\name{get_tsne_embedding}
\alias{get_tsne_embedding}
\title{Compute an embedding of the multi-omics matrix using t-SNE}
\usage{
get_tsne_embedding(matrix, embedding_size, perplexity, max_iter, num_threads)
}
\arguments{
\item{matrix}{A squared numeric matrix with values in the range 0-1, with samples on columns}

\item{embedding_size}{Size of the output embedding}

\item{perplexity}{Perplexity parameter}

\item{max_iter}{Number of iterations}

\item{num_threads}{Number of threads for parallelization. It has to be positive integer or 0, if all available cores must be used. If it is equal to 1, no parallelization is not performed}
}
\value{
An embedding of the input Matrix: on the columns there are the samples, the number of rows are specified by embedding_size
}
\description{
Compute an embedding of the multi-omics matrix using t-SNE
}
