% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_star_net.R
\name{k_star_net}
\alias{k_star_net}
\title{Build omic similarity network}
\usage{
k_star_net(
  matrix,
  distFun = "Euclidean",
  sparsity = 1,
  knn = 25,
  k_star = TRUE,
  cores = 1,
  MAX_ASSOC = Inf
)
}
\arguments{
\item{matrix}{a numeric matrix representing a particular omics, features are on rows and samples are on columns}

\item{distFun}{a distance function chosen among: "Euclidean", "Manhattan", "Cosine"}

\item{sparsity}{a positive real value guiding the choice of the number of neighbors for each node based on the distribution of distances of its neighbors}

\item{knn}{an integer, it is the maximum number of neighbors to be considered as candidate neighbors for each node}

\item{k_star}{if TRUE  the number of neighbors id dynamically determined for each node using kstar algorithm ( https://papers.nips.cc/paper/2016/file/2c6ae45a3e88aee548c0714fad7f8269-Paper.pdf),
if FALSE k closest neighbors are chosen for each node}

\item{cores}{number of threads for parallelization. It has to be positive integer. If it is equal to 1, no parallelization is not performed}

\item{MAX_ASSOC}{number of maximum incoming edges that a node can have}
}
\value{
a tibble containing the edge list of output network with the following columns: source, dest, weight.
Source and dest represent the edges and contain node ids, weight contains edge weights representing nodes similarity.
}
\description{
Build omic similarity network
}
