% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_omics.R
\name{normalize_omics}
\alias{normalize_omics}
\title{Center and scale an omic matrix along its columns}
\usage{
normalize_omics(matrix)
}
\arguments{
\item{matrix}{a numeric matrix representing a particular omics, features are on rows and samples are on columns}
}
\value{
a matrix center and scaled along its columns
}
\description{
Center and scale an omic matrix along its columns
}
