% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_omics.R
\name{normalize_omics}
\alias{normalize_omics}
\title{Normalize matrix using ???}
\usage{
normalize_omics(mtx)
}
\arguments{
\item{mtx}{a omics matrix}
}
\value{
normalized omics matrix
}
\description{
Normalize matrix using ???
}
