% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Create jump matrix}
\usage{
plot_net(edgeList, nodes_anno, name, html = FALSE)
}
\arguments{
\item{edgeList}{dataframe representing a graph}

\item{nodes_anno}{nodes annotation}

\item{name}{name of plot}

\item{html}{bool if you want to save the plot in a html file}
}
\description{
Create jump matrix
}
