% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot umap}
\usage{
plot_umap(
  mat,
  anno_df,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  title = ""
)
}
\arguments{
\item{mat}{matrix with samples on columns}

\item{anno_df}{annotation dataframe}

\item{id_anno_color}{color of identification of annotation dataframe}

\item{id_anno_shape}{shape of identification of annotation dataframe}

\item{id}{identification of annotation dataframe}
}
\value{
plot umap
}
\description{
Plot umap
}
