% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot umap}
\usage{
plot_umap(
  matrix,
  nodes_anno,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  title = ""
)
}
\arguments{
\item{matrix}{A Matrix with samples on rows that has to be plotted}

\item{nodes_anno}{Annotation DataFrame with all the information for each sample}

\item{id_name}{String for identification of sample in nodes_anno}

\item{id_anno_color}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific color for each case}

\item{id_anno_shape}{String for the column necessary for distinguish cases in the nodes_anno, it will be used for giving a specific shape for each case}
}
\value{
A plot
}
\description{
Plot umap
}
