% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_db.R
\name{plot_umap_db}
\alias{plot_umap_db}
\title{Create jump matrix}
\usage{
plot_umap_db(mat, nodes_anno, name = NA, type = "data", k = 1)
}
\arguments{
\item{mat}{matrix}

\item{nodes_anno}{nodes annotation}

\item{name}{name of plot}

\item{type}{type}

\item{k}{k}
}
\description{
Create jump matrix
}
