% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_db.R
\name{plot_umap_db}
\alias{plot_umap_db}
\title{Plot umap db}
\usage{
plot_umap_db(matrix, type = "data", k = 1)
}
\arguments{
\item{matrix}{A Matrix with samples on rows that has to be plotted}

\item{type}{Type of matrix, used as input for umap, see umap documentation}

\item{k}{Reachability distance, dbscan}
}
\description{
Plot umap db
}
