% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_multiplex_network.R
\name{prune_multiplex_network}
\alias{prune_multiplex_network}
\title{Remove extra weigth}
\usage{
prune_multiplex_network(network, weight_remv = NULL)
}
\arguments{
\item{network}{a network with three columns}

\item{weight_remv}{the weight to be removed in the network}
}
\value{
network with weigths strictly less than weight_remv
}
\description{
Remove extra weigth
}
