% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zeros_cols.R
\name{remove_zeros_cols}
\alias{remove_zeros_cols}
\title{Remove zero columns from a matrix}
\usage{
remove_zeros_cols(matrix)
}
\arguments{
\item{matrix}{a numeric matrix}
}
\value{
a matrix of the same type of \emph{matrix} with columns containing all zero removed
}
\description{
Remove zero columns from a matrix
}
