% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zeros_cols.R
\name{remove_zeros_cols}
\alias{remove_zeros_cols}
\title{Create jump matrix}
\usage{
remove_zeros_cols(mtx)
}
\arguments{
\item{mtx}{a matrix}
}
\value{
the same matrix as input without the zeros columns
}
\description{
Create jump matrix
}
